e2e: Eliminate the util/crud.go
Split out the drpolicy function to drpolicy.go and rename the file to
namespace.go. Now we can understand better the structure of the util
package.
